+++ recipes/common.rb
@@ -118,16 +118,8 @@ db_user = node['openstack']['db']['network']['username']
 db_pass = get_password 'db', 'neutron'
 sql_connection = db_uri('network', db_user, db_pass)
 
-api_endpoint = endpoint 'network-api'
+network_api_bind = endpoint 'network-api-bind'
 service_pass = get_password 'service', 'openstack-network'
-
-if node['openstack']['network']['api']['bind_interface'].nil?
-  bind_address = api_endpoint.host
-  bind_port = api_endpoint.port
-else
-  bind_address = address_for node['openstack']['network']['api']['bind_interface']
-  bind_port = node['openstack']['network']['api']['bind_port']
-end
 
 platform_options['neutron_client_packages'].each do |pkg|
   package pkg do
@@ -154,8 +146,8 @@ template '/etc/neutron/neutron.conf' do
   group node['openstack']['network']['platform']['group']
   mode   00644
   variables(
-    bind_address: bind_address,
-    bind_port: bind_port,
+    bind_address: network_api_bind.host,
+    bind_port: network_api_bind.port,
     rabbit_hosts: rabbit_hosts,
     mq_service_type: mq_service_type,
     mq_password: mq_password,